# Optimize output vector adapter write